Add files via upload
This is a small number guessing game I made using Python and the random module. The program picks a random number between 1 and 100, and the player has to guess it within 5 tries. It’s nothing too advanced, but it’s one of my first projects as I’m learning to code. I’m sharing it here as part of my journey to keep track of my progress.😊